requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.2.0
asyncio==3.4.3
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
d6tcollect==1.0.6
d6tflow==0.2.6
docker==7.0.0
docutils==0.20.1
fastapi==0.109.0
Flask==3.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lockfile==0.12.2
luigi==3.5.0
markdown2==2.4.10
MarkupSafe==2.1.3
nicegui==1.4.11
numpy==1.26.3
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pscript==0.7.7
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-multipart==0.0.6
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
replicate==0.22.0
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
simple-websocket==1.0.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
